renamed selfcare client
