+++ packages/api-clients/src/index.ts
@@ -7,4 +7,4 @@ export * as catalogApi from "./catalogApi.js";
 export * as purposeApi from "./generated/purposeApi.js";
 export * as selfcareV2ClientApi from "./generated/selfcareV2ClientApi.js";
 export * as tenantApi from "./generated/tenantApi.js";
-export * from "./client.js";
+export * from "./selfcareClients.js";
